# movie_time_backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Movie Explorer API", "description": "\n    ## Movie Explorer Backend API\n    \n    A comprehensive API for exploring movies, actors, directors, and genres.\n    \n    ### Features\n    - Full CRUD operations for movies\n    - Query movies by genre, actor, director, or release year\n    - Explore actors and directors\n    - Browse genres\n    \n    ### Response Format\n    All responses follow a consistent format:\n    ```json\n    {\n        \"success\": true/false,\n        \"message\": \"Human readable message\",\n        \"data\": {}\n    }\n    ```\n    ", "version": "1.0.0"}, "paths": {"/movies/details": {"get": {"tags": ["Movies"], "summary": "Get all movies with full details", "description": "Retrieve a list of all movies with title, release year, poster URL, director name, actors, and genres.", "operationId": "get_movies_details_movies_details_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Get Movies Details Movies Details Get"}}}}}}}, "/movies/summary": {"get": {"tags": ["Movies"], "summary": "Get simplified movie list", "description": "Retrieve a simplified list of movies with title, director name, actor names, and rating.", "operationId": "get_movies_summary_movies_summary_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Get Movies Summary Movies Summary Get"}}}}}}}, "/movies": {"get": {"tags": ["Movies"], "summary": "Get all movies", "description": "Retrieve a list of all movies. Can be filtered by genre_id, actor_id, director_id, and release_year.", "operationId": "get_movies_movies_get", "parameters": [{"name": "genre_id", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "Filter by genre ID", "title": "Genre Id"}, "description": "Filter by genre ID"}, {"name": "actor_id", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "Filter by actor ID", "title": "Actor Id"}, "description": "Filter by actor ID"}, {"name": "director_id", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "Filter by director ID", "title": "Director Id"}, "description": "Filter by director ID"}, {"name": "release_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "description": "Filter by release year", "title": "Release Year"}, "description": "Filter by release year"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Get Movies Movies Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Movies"], "summary": "Create a new movie", "description": "Create a new movie with the provided details.", "operationId": "create_movie_movies_post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MovieCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Create Movie Movies Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/movies/{movie_id}": {"get": {"tags": ["Movies"], "summary": "Get movie by ID", "description": "Retrieve a specific movie by its ID.", "operationId": "get_movie_movies__movie_id__get", "parameters": [{"name": "movie_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Movie Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Get Movie Movies  Movie Id  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Movies"], "summary": "Update a movie", "description": "Update an existing movie with the provided details.", "operationId": "update_movie_movies__movie_id__put", "parameters": [{"name": "movie_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Movie Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MovieUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Update Movie Movies  Movie Id  Put"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Movies"], "summary": "Delete a movie", "description": "Delete a movie by its ID.", "operationId": "delete_movie_movies__movie_id__delete", "parameters": [{"name": "movie_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Movie Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Delete Movie Movies  Movie Id  Delete"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/actors": {"get": {"tags": ["Actors"], "summary": "Get all actors", "description": "Retrieve a list of all actors. Can be filtered by movie_id or genre_id.", "operationId": "get_actors_actors_get", "parameters": [{"name": "movie_id", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "Filter by movie ID", "title": "Movie Id"}, "description": "Filter by movie ID"}, {"name": "genre_id", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "Filter by genre ID (actors in movies of this genre)", "title": "Genre Id"}, "description": "Filter by genre ID (actors in movies of this genre)"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Get Actors Actors Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Actors"], "summary": "Create a new actor", "description": "Create a new actor with the provided details.", "operationId": "create_actor_actors_post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActorCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Create Actor Actors Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/actors/{actor_id}": {"get": {"tags": ["Actors"], "summary": "Get actor by ID", "description": "Retrieve a specific actor by their ID.", "operationId": "get_actor_actors__actor_id__get", "parameters": [{"name": "actor_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Actor Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Get Actor Actors  Actor Id  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/directors": {"get": {"tags": ["Directors"], "summary": "Get all directors", "description": "Retrieve a list of all directors.", "operationId": "get_directors_directors_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Get Directors Directors Get"}}}}}}, "post": {"tags": ["Directors"], "summary": "Create a new director", "description": "Create a new director with the provided details.", "operationId": "create_director_directors_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/DirectorCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Create Director Directors Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/directors/{director_id}": {"get": {"tags": ["Directors"], "summary": "Get director by ID", "description": "Retrieve a specific director by their ID.", "operationId": "get_director_directors__director_id__get", "parameters": [{"name": "director_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Director Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Get Director Directors  Director Id  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/genres": {"get": {"tags": ["Genres"], "summary": "Get all genres", "description": "Retrieve a list of all available genres.", "operationId": "get_genres_genres_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Get Genres Genres Get"}}}}}}, "post": {"tags": ["Genres"], "summary": "Create a new genre", "description": "Create a new genre with the provided details.", "operationId": "create_genre_genres_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenreCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Create Genre Genres Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/genres/{genre_id}": {"get": {"tags": ["Genres"], "summary": "Get genre by ID", "description": "Retrieve a specific genre by its ID.", "operationId": "get_genre_genres__genre_id__get", "parameters": [{"name": "genre_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Genre Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Get Genre Genres  Genre Id  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"tags": ["Health"], "summary": "Root", "description": "Root endpoint - health check.", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/health": {"get": {"tags": ["Health"], "summary": "Health Check", "description": "Health check endpoint.", "operationId": "health_check_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"ActorCreate": {"properties": {"name": {"type": "string", "maxLength": 200, "minLength": 1, "title": "Name", "description": "Actor name"}, "bio": {"type": "string", "maxLength": 2000, "minLength": 1, "title": "Bio", "description": "Actor biography"}, "movie_ids": {"items": {"type": "string"}, "type": "array", "title": "Movie Ids", "description": "List of movie IDs the actor appeared in", "default": []}}, "type": "object", "required": ["name", "bio"], "title": "ActorCreate", "description": "Schema for creating a new actor.", "examples": [{"bio": "American actor and film producer known for his work in biopics and period films.", "movie_ids": [], "name": "Leonardo DiCaprio"}]}, "DirectorCreate": {"properties": {"name": {"type": "string", "maxLength": 200, "minLength": 1, "title": "Name", "description": "Director name"}, "bio": {"type": "string", "maxLength": 2000, "minLength": 1, "title": "Bio", "description": "Director biography"}, "movie_ids": {"items": {"type": "string"}, "type": "array", "title": "Movie Ids", "description": "List of movie IDs directed", "default": []}}, "type": "object", "required": ["name", "bio"], "title": "DirectorCreate", "description": "Schema for creating a new director.", "examples": [{"bio": "British-American filmmaker known for his distinctive narrative style and complex storytelling.", "movie_ids": [], "name": "Christopher Nolan"}]}, "GenreCreate": {"properties": {"name": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Name", "description": "Genre name"}, "description": {"type": "string", "maxLength": 500, "minLength": 1, "title": "Description", "description": "Genre description"}}, "type": "object", "required": ["name", "description"], "title": "GenreCreate", "description": "Schema for creating a new genre.", "examples": [{"description": "Movies with exciting action sequences, stunts, and physical feats.", "name": "Action"}]}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "MovieCreate": {"properties": {"title": {"type": "string", "maxLength": 300, "minLength": 1, "title": "Title", "description": "Movie title"}, "release_year": {"type": "integer", "maximum": 2100.0, "minimum": 1888.0, "title": "Release Year", "description": "Release year"}, "director_id": {"type": "string", "title": "Director Id", "description": "Director's ObjectId"}, "actor_ids": {"items": {"type": "string"}, "type": "array", "title": "Actor Ids", "description": "List of actor ObjectIds", "default": []}, "genre_ids": {"items": {"type": "string"}, "type": "array", "title": "Genre Ids", "description": "List of genre ObjectIds", "default": []}, "rating": {"type": "number", "maximum": 10.0, "minimum": 0.0, "title": "Rating", "description": "Movie rating from 0 to 10"}, "reviews": {"items": {"$ref": "#/components/schemas/Review"}, "type": "array", "title": "Reviews", "description": "List of reviews", "default": []}}, "type": "object", "required": ["title", "release_year", "director_id", "rating"], "title": "MovieCreate", "description": "Schema for creating a new movie.", "examples": [{"actor_ids": ["507f1f77bcf86cd799439012", "507f1f77bcf86cd799439013"], "director_id": "507f1f77bcf86cd799439011", "genre_ids": ["507f1f77bcf86cd799439014", "507f1f77bcf86cd799439015"], "rating": 8.8, "release_year": 2010, "reviews": [{"comment": "Mind-bending plot with amazing visuals!", "rating": 9.0, "user": "moviefan123"}], "title": "Inception"}]}, "MovieUpdate": {"properties": {"title": {"anyOf": [{"type": "string", "maxLength": 300, "minLength": 1}, {"type": "null"}], "title": "Title", "description": "Movie title"}, "release_year": {"anyOf": [{"type": "integer", "maximum": 2100.0, "minimum": 1888.0}, {"type": "null"}], "title": "Release Year", "description": "Release year"}, "director_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Director Id", "description": "Director's ObjectId"}, "actor_ids": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Actor Ids", "description": "List of actor ObjectIds"}, "genre_ids": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Genre Ids", "description": "List of genre ObjectIds"}, "rating": {"anyOf": [{"type": "number", "maximum": 10.0, "minimum": 0.0}, {"type": "null"}], "title": "Rating", "description": "Movie rating from 0 to 10"}, "reviews": {"anyOf": [{"items": {"$ref": "#/components/schemas/Review"}, "type": "array"}, {"type": "null"}], "title": "Reviews", "description": "List of reviews"}, "poster_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Poster Url", "description": "Movie poster URL"}}, "type": "object", "title": "MovieUpdate", "description": "Schema for updating a movie."}, "Review": {"properties": {"user": {"type": "string", "maxLength": 100, "minLength": 1, "title": "User", "description": "Username of reviewer"}, "comment": {"type": "string", "maxLength": 1000, "minLength": 1, "title": "Comment", "description": "Review comment"}, "rating": {"type": "number", "maximum": 10.0, "minimum": 0.0, "title": "Rating", "description": "Rating from 0 to 10"}}, "type": "object", "required": ["user", "comment", "rating"], "title": "Review", "description": "Schema for movie review.", "examples": [{"comment": "An absolute masterpiece! The cinematography is stunning.", "rating": 9.5, "user": "moviefan123"}]}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}